+++ src/Dash/src/Dash/Application/ApplicationServices.cs
@@ -2,7 +2,8 @@
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 
 using Dash.Application.Default;
-using Dash.Application.PreprocessingSteps;
+using Dash.PreprocessingSteps;
+using Dash.PreprocessingSteps.Default;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Dash.Application
