+++ src/Dash/src/Dash/Dash.csproj
@@ -64,4 +64,8 @@
     </None>
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="PreprocessingSteps\Default\" />
+  </ItemGroup>
+
 </Project>
