+++ src/Dash/test/Dash.Tests/Application/Default/DashOptionsValidatorTests.cs
@@ -4,8 +4,8 @@
 using System.IO.Abstractions.TestingHelpers;
 using System.Threading.Tasks;
 using Dash.Application;
-using Dash.Application.PreprocessingSteps;
 using Dash.Common;
+using Dash.PreprocessingSteps.Default;
 using FluentAssertions;
 using Microsoft.Extensions.Options;
 using NSubstitute;
